Dark background on messages
This adds a dark background on messages when you hover over them, similar to Discord. This is in preparation for right clicking messages so you know what you're clicking.

This is worth it's own commit cause I had to redo a lot of the way messages were adding to html.