Fixed a serious vulnerability
There was a serious remote code execution vulnerability that allowed anybody to execute commands on the machine of everybody they came into contact with, who was using Livebot. The 'hacker' would not have to use livebot to run this attack.

The exact vulnerabilty will not be spread to anybody for the time being