Multiple styling fixes, worked on user settings
The proper way of doing ellipsis on the text (adding ... if it's too long) has been added. Idk why I was trying to do it with js before, it's a css thing...

The user settings menu has also been worked on some more. Changing your username is currently functional. If there is an error setting the username (too short, changing too often, etc) then the textbox highlights red, and an error appears in console. Will try to add a displaying message tomrrow

Also, now when you select a channel, the placeholder in the message box displays 'Message #channel-name'. Easy addition, but here it is