Some minor fixes, began work on member menu
There have been some styling fixes, some other small js fixes, but most importantly, is that the member menu is beginning to be worked on

The menu so far only shows the custom presence, username, tag, and icon, but more will come soon. The emoji only shows if it's not a custom emoji, that will be fixed soon as well.