Every guild now has default channel
Before, when you clicked on a server, it would open the last channel you had open in there, but that would require you to have opened that channel already in livebot. Now it does the same thing, but if there is no channel saved for it to open, it finds the first channel in the list that the bot has permission to view, and opens that!

Other small fixes