No longer using localstorage
The localstorage has been exchanged for a json file, and it is stored in ~/json/logins.json (which has also been added to .gitignore, so don't worry about tokens being uploaded)

Electron has been upgraded to latest, 13.1.6. This meant that custom-electron-titlebar was no longer operational, so I made up a custom titlebar. It looks great, acts great, and no longer changes colour when you tab out of the window (just like discord's), so it blends in better to the background now.

If the login issues continue to happen, all a user needs to do is send the logins.json file to staff (probably privately, since the token is still stored in there), and we can do some proper debugging. Also to fix login problems, all that needs to happen is to delete the logins.json file. No more digging around in console if you're not familliar with it.