Changed dir to active directory on startup
When LiveBot was used as an executable, by default, it would think the active directory was the main folder (livebot-win32-x64). Now, upon startup, livebot sets the active directory to wherever all the actual code is. This means it should actually work.

New releases will be coming because of this mistake.