Improved error handling, modified scripts manager
Tested the embedbuilder and script manager. I made slight modifications to the scripts to use the 'legacy' exports.start, which allows the script to contain exports.info as was originally intended. I also then added this info to the scripts page if it exists. We still support scripts that are just module.exports, but they will not show additional information